Extrait la connexion à Grist
